+++ tablet_qt/core/networkmanager.h
@@ -167,6 +167,7 @@ protected:
     void queryFailClearingMoveOffFlag(const QString& tablename);
     bool clearMoveOffTabletFlag(const QString& tablename);
     bool pruneDeadBlobs();
+    bool serverSupportsValidatePatients() const;
     bool serverSupportsOneStepUpload() const;
     bool shouldUseOneStepUpload() const;
     void uploadOneStep();